refactor(auth): flatten single-file leaf directories
Collapse 18 single-file wrapper directories in src/modules/auth into
their parents. Role suffixes (.helper/.mapper/.schema/.strategy/.policy/
.entity/.config/.adapter/.repository) already carry the categorisation
the folders duplicated, so no semantic signal is lost.

Flattened clusters: infrastructure/session/{helpers,mappers,schemas,
strategies}, infrastructure/crypto/{adapters,config}, presentation/authn/
components/shared/{controls,feedback,links}, presentation/session/
{actions,features,guards}, domain/session/policies/{lifecycle,security}
(authorization/ kept), plus lone dirs infrastructure/logging,
persistence/auth-user/repositories, application/auth-user/errors,
domain/session/types.

Mechanical move (git mv) + alias-path rewrite across 37 referrers;
multi-file directories (adapters/config/services/types/authorization/...)
left intact. Implements fix #4 from the 2026-06-12 auth-structure
assessment.

Verified: pnpm check:fast (lint + tsc + typegen) and 226/226 unit tests.
