+++ src/modules/auth/__tests__/integration/auth-user.repository.final-gate.integration.test.ts
@@ -2,7 +2,7 @@ import { users } from "@database/schema/users";
 import { eq } from "drizzle-orm";
 import { afterEach, describe, expect, it } from "vitest";
 import type { AuthUserCreateDto } from "@/modules/auth/application/auth-user/dtos/requests/auth-user-create.dto";
-import { AuthUserRepository } from "@/modules/auth/infrastructure/persistence/auth-user/repositories/auth-user.repository";
+import { AuthUserRepository } from "@/modules/auth/infrastructure/persistence/auth-user/auth-user.repository";
 import { getAppDb } from "@/server/db/db.connection";
 import { logger } from "@/shared/telemetry/logging/infrastructure/logging.client";
