+++ src/modules/auth/application/auth-user/auth-error.factory.ts
