+++ src/modules/auth/application/session/contracts/session-service.contract.ts
@@ -2,7 +2,7 @@ import type { ReadSessionOutcomeDto } from "@/modules/auth/application/session/d
 import type { SessionPrincipalDto } from "@/modules/auth/application/session/dtos/responses/session-principal.dto";
 import type { SessionVerificationDto } from "@/modules/auth/application/session/dtos/responses/session-verification.dto";
 import type { UpdateSessionOutcomeDto } from "@/modules/auth/application/session/dtos/responses/update-session-outcome.dto";
-import type { TerminateSessionReason } from "@/modules/auth/domain/session/policies/lifecycle/evaluate-session-lifecycle.policy";
+import type { TerminateSessionReason } from "@/modules/auth/domain/session/policies/evaluate-session-lifecycle.policy";
 import type { AppError } from "@/shared/core/errors/core/app-error.entity";
 import type { Result } from "@/shared/core/result/result.dto";
