+++ src/modules/auth/domain/session/auth-request-authorization.output.ts
