+++ src/modules/auth/domain/session/policies/evaluate-session-lifecycle.policy.ts
