+++ src/modules/auth/infrastructure/composition/factories/crypto/password-hasher.factory.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import type { PasswordHasherContract } from "@/modules/auth/application/auth-user/contracts/services/password-hasher.contract";
-import { PasswordHasherAdapter } from "@/modules/auth/infrastructure/crypto/adapters/password-hasher.adapter";
-import { getAuthCryptoConfig } from "@/modules/auth/infrastructure/crypto/config/auth-crypto.config";
+import { getAuthCryptoConfig } from "@/modules/auth/infrastructure/crypto/auth-crypto.config";
+import { PasswordHasherAdapter } from "@/modules/auth/infrastructure/crypto/password-hasher.adapter";
 import { BcryptPasswordService } from "@/modules/auth/infrastructure/crypto/services/bcrypt-password.service";
 
 /**
