+++ src/modules/auth/infrastructure/crypto/auth-crypto.config.ts
