+++ src/modules/auth/infrastructure/persistence/auth-user/auth-user.repository.ts
