+++ src/modules/auth/infrastructure/session/services/session-token.service.ts
@@ -8,8 +8,8 @@ import type { SessionTokenClaimsDto } from "@/modules/auth/application/session/d
 import { toSessionTokenClaimsDto } from "@/modules/auth/application/session/mappers/to-session-token-claims-dto.mapper";
 import { SESSION_DURATION_SEC } from "@/modules/auth/domain/shared/constants/session-config.constants";
 import { SESSION_TOKEN_CLOCK_TOLERANCE_SEC } from "@/modules/auth/infrastructure/session/config/session-token.constants";
-import { jwtToSessionTokenClaimsDto } from "@/modules/auth/infrastructure/session/mappers/jwt-to-session-token-claims-dto.mapper";
-import { SessionTokenClaimsSchema } from "@/modules/auth/infrastructure/session/schemas/session-token-claims.schema";
+import { jwtToSessionTokenClaimsDto } from "@/modules/auth/infrastructure/session/jwt-to-session-token-claims-dto.mapper";
+import { SessionTokenClaimsSchema } from "@/modules/auth/infrastructure/session/session-token-claims.schema";
 import type { AppError } from "@/shared/core/errors/core/app-error.entity";
 import { APP_ERROR_KEYS } from "@/shared/core/errors/core/catalog/app-error.registry";
 import { makeAppError } from "@/shared/core/errors/core/factories/app-error.factory";
