+++ src/modules/auth/infrastructure/session/services/session.service.ts
@@ -10,7 +10,7 @@ import type { SessionVerificationDto } from "@/modules/auth/application/session/
 import type { UpdateSessionOutcomeDto } from "@/modules/auth/application/session/dtos/responses/update-session-outcome.dto";
 import { ReadSessionUseCase } from "@/modules/auth/application/session/queries/read-session.use-case";
 import { RequireSessionUseCase } from "@/modules/auth/application/session/queries/require-session.use-case";
-import type { TerminateSessionReason } from "@/modules/auth/domain/session/policies/lifecycle/evaluate-session-lifecycle.policy";
+import type { TerminateSessionReason } from "@/modules/auth/domain/session/policies/evaluate-session-lifecycle.policy";
 import type { AppError } from "@/shared/core/errors/core/app-error.entity";
 import { Err, Ok } from "@/shared/core/result/result";
 import type { Result } from "@/shared/core/result/result.dto";
