+++ src/modules/invoices/presentation/actions/create-invoice.action.ts
@@ -1,6 +1,6 @@
 "use server";
 import { revalidatePath } from "next/cache";
-import { requireSession } from "@/modules/auth/presentation/session/guards/session-access.guard";
+import { requireSession } from "@/modules/auth/presentation/session/session-access.guard";
 import { InvoiceService } from "@/modules/invoices/application/services/invoice.service";
 import { toInvoiceErrorMessage } from "@/modules/invoices/application/utils/error-messages";
 import { INVOICE_MSG } from "@/modules/invoices/domain/i18n/invoice-messages";
