+++ src/modules/users/__tests__/unit/presentation/actions/create-user.action.test.ts
@@ -4,7 +4,7 @@ import {
 	TEST_USERNAME,
 } from "@test-support/fixtures/user.fixtures";
 import { beforeEach, describe, expect, it, type Mock, vi } from "vitest";
-import { requireAdmin } from "@/modules/auth/presentation/session/guards/session-access.guard";
+import { requireAdmin } from "@/modules/auth/presentation/session/session-access.guard";
 import {
 	USER_ERROR_MESSAGES,
 	USER_SUCCESS_MESSAGES,
@@ -28,16 +28,13 @@ import { validateForm } from "@/shared/forms/server/validate-form";
 vi.mock("@/shared/forms/server/validate-form");
 vi.mock("@/modules/users/infrastructure/factories/user-service.factory");
 vi.mock("@/server/db/db.connection");
-vi.mock(
-	"@/modules/auth/presentation/session/guards/session-access.guard",
-	() => ({
-		requireAdmin: vi.fn().mockResolvedValue({
-			isAuthorized: true,
-			role: "ADMIN",
-			userId: "admin-1",
-		}),
+vi.mock("@/modules/auth/presentation/session/session-access.guard", () => ({
+	requireAdmin: vi.fn().mockResolvedValue({
+		isAuthorized: true,
+		role: "ADMIN",
+		userId: "admin-1",
 	}),
-);
+}));
 vi.mock(
 	"@/shared/forms/logic/factories/form-result.factory",
 	async (importOriginal) => {
