+++ src/modules/users/presentation/actions/read-user.action.ts
@@ -1,5 +1,5 @@
 "use server";
-import { requireAdmin } from "@/modules/auth/presentation/session/guards/session-access.guard";
+import { requireAdmin } from "@/modules/auth/presentation/session/session-access.guard";
 import type { UserDto } from "@/modules/users/application/dtos/user.dto";
 import { toUserId } from "@/modules/users/domain/user-id.mappers";
 import { createUserService } from "@/modules/users/infrastructure/factories/user-service.factory";
